refactor: Move the useGeneralActions hook to Hooks folder
